requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.52
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.7.0
transformers==4.36.0
tokenizers==0.15.0
huggingface-hub==0.23.0
sentence-transformers==2.6.0
accelerate==0.20.3
python-dotenv==1.0.1
google-cloud-bigquery==3.20.0
google-cloud-bigquery-storage==2.24.0
db-dtypes==1.1.1
pandas==2.2.0
numpy==1.24.3
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
python-docx==1.1.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
langchain-huggingface==0.2.0
